+++ plugins/provisioners/docker/cap/linux/docker_daemon_running.rb
@@ -1,0 +1,13 @@
+module VagrantPlugins
+  module Docker
+    module Cap
+      module Linux
+        module DockerDaemonRunning
+          def self.docker_daemon_running(machine)
+            machine.communicate.test("test -f /var/run/docker/pid")
+          end
+        end
+      end
+    end
+  end
+end